-created menu bar -added basic plant activity -added functionality to swap between activities -deleted MainActivity.kt, search.png and menu_main.xml as no longer needed. -
