+++ app/src/main/AndroidManifest.xml
@@ -26,11 +26,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.MainActivity"
-            android:exported="false"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.Plantico.NoActionBar"/>
-        <activity
             android:name=".ui.search.SearchableActivity"
             android:exported="false"
             android:label="@string/app_name"
